// apps/frontend/src/shared/api/globals.d.ts
export interface Organization {
  /**
   * 组织 ID
   */
  id: string;
  /**
   * 组织名
   */
  name: string;
  /**
   * 父组织 ID,根组织为 null
   */
  parentId: string | null;
  /**
   * 创建时间(ISO 8601)
   */
  createdAt: string;
  /**
   * 更新时间(ISO 8601)
   */
  updatedAt: string;
}